remove main.py dependence
