+++ burnman/__init__.py
@@ -185,7 +185,6 @@ from .combinedmineral import CombinedMineral
 from .mineral_helpers import *
 
 # high level functions
-from .main import *
 from .model import Model
 
 # mineral library
