+++ burnman/main.py
@@ -49,114 +49,3 @@ def velocities_from_rock(rock, pressures, temperatures, averaging_scheme=averagi
         ['rho', 'v_p', 'v_s', 'v_phi', 'K_S', 'G'], pressures, temperatures)
     rock.set_averaging_scheme(old_averaging_scheme)
     return rho, vp, vs, vphi, K, G
-
-
-def compare_l2(depth, calc, obs):
-    """
-    PUT IN TOOLS
-    Computes the L2 norm for N profiles at a time (assumed to be linear between points).
-
-    .. math:: math does not work yet...
-       \sum_{i=1}^{\\infty} x_{i}
-
-    :type depths: array of float
-    :param depths: depths. :math:`[m]`
-    :type calc: list of arrays of float
-    :param calc: N arrays calculated values, e.g. [mat_vs,mat_vphi]
-    :type obs: list of arrays of float
-    :param obs: N arrays of values (observed or calculated) to compare to , e.g. [seis_vs, seis_vphi]
-
-    :returns: array of L2 norms of length N
-    :rtype: array of floats
-    """
-    err = []
-    for l in range(len(calc)):
-        err.append(l2(depth, calc[l], obs[l]))
-
-    return err
-
-
-def compare_chifactor(calc, obs):
-    """
-    PUT IN TOOLS
-    Computes the chi factor for N profiles at a time. Assumes a 1% a priori uncertainty on the seismic model.
-
-
-    :type calc: list of arrays of float
-    :param calc: N arrays calculated values, e.g. [mat_vs,mat_vphi]
-    :type obs: list of arrays of float
-    :param obs: N arrays of values (observed or calculated) to compare to , e.g. [seis_vs, seis_vphi]
-
-    :returns: error array of length N
-    :rtype: array of floats
-    """
-    err = []
-    for l in range(len(calc)):
-        err.append(chi_factor(calc[l], obs[l]))
-
-    return err
-
-
-def l2(x, funca, funcb):
-    """
-    PUT IN TOOLS
-    Computes the L2 norm for one profile(assumed to be linear between points).
-
-    :type x: array of float
-    :param x: depths :math:`[m]`.
-    :type funca: list of arrays of float
-    :param funca: array calculated values
-    :type funcb: list of arrays of float
-    :param funcb: array of values (observed or calculated) to compare to
-
-    :returns: L2 norm
-    :rtype: array of floats
-    """
-    diff = np.array(funca - funcb)
-    diff = diff * diff
-    return integrate.trapz(diff, x)
-
-
-def nrmse(x, funca, funcb):
-    """
-    PUT IN TOOLS
-    Normalized root mean square error for one profile
-    :type x: array of float
-    :param x: depths in m.
-    :type funca: list of arrays of float
-    :param funca: array calculated values
-    :type funcb: list of arrays of float
-    :param funcb: array of values (observed or calculated) to compare to
-
-    :returns: RMS error
-    :rtype: array of floats
-
-    """
-    diff = np.array(funca - funcb)
-    diff = diff * diff
-    rmse = np.sqrt(np.sum(diff) / x)
-    nrmse = rmse / (np.max(funca) - np.min(funca))
-    return nrmse
-
-
-def chi_factor(calc, obs):
-    """
-    PUT IN TOOLS
-    :math:`\\chi` factor for one profile assuming 1% uncertainty on the reference model (obs)
-    :type calc: list of arrays of float
-    :param calc: array calculated values
-    :type obs: list of arrays of float
-    :param obs: array of reference values to compare to
-
-    :returns: :math:`\\chi` factor
-    :rtype: array of floats
-
-    """
-
-    err = np.empty_like(calc)
-    for i in range(len(calc)):
-        err[i] = pow((calc[i] - obs[i]) / (0.01 * np.mean(obs)), 2.)
-
-    err_tot = np.sum(err) / len(err)
-
-    return err_tot
